# Update parquet go