--- a/cmd/tempo-cli/cmd-rewrite-blocks_test.go
+++ b/cmd/tempo-cli/cmd-rewrite-blocks_test.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bytes"
 	"context"
+	"errors"
 	"io"
 	"sort"
 	"strings"
@@ -135,7 +136,7 @@ func getAllTraceIDs(t *testing.T, dir string, tenant string) []string {
 		for read := int64(0); read < r.NumRows(); {
 			rows := make([]parquet.Row, r.NumRows())
 			n, err := r.ReadRows(rows)
-			if err != io.EOF {
+			if !errors.Is(err, io.EOF) {
 				require.NoError(t, err)
 			}
 			require.Greater(t, n, 0)