--- a/tempodb/encoding/vparquet2/wal_block.go
+++ b/tempodb/encoding/vparquet2/wal_block.go
@@ -743,7 +743,7 @@ func (i *rowIterator) Next(context.Context) (common.ID, parquet.Row, error) {
 
 	rows := []parquet.Row{completeBlockRowPool.Get()}
 	_, err = i.reader.ReadRows(rows)
-	if err != nil {
+	if err != nil && err != io.EOF {
 		return nil, nil, err
 	}