--- a/tempodb/encoding/vparquet4/wal_block.go
+++ b/tempodb/encoding/vparquet4/wal_block.go
@@ -910,7 +910,7 @@ func (i *rowIterator) Next(context.Context) (common.ID, parquet.Row, error) {
 
 	rows := []parquet.Row{completeBlockRowPool.Get()}
 	_, err = i.reader.ReadRows(rows)
-	if err != nil && err != io.EOF {
+	if err != nil && !errors.Is(err, io.EOF) {
 		return nil, nil, err
 	}