… go.sum
… vendor/github.com/parquet-go/parquet-go/.gitignore
… vendor/github.com/parquet-go/parquet-go/CODEOWNERS
… vendor/github.com/parquet-go/parquet-go/README.md
… vendor/github.com/parquet-go/parquet-go/allocator.go
… vendor/github.com/parquet-go/parquet-go/bloom.go
… vendor/github.com/parquet-go/parquet-go/bloom/xxhash/sum64uint_amd64.s
… vendor/github.com/parquet-go/parquet-go/buffer.go
… vendor/github.com/parquet-go/parquet-go/column.go
… vendor/github.com/parquet-go/parquet-go/column_buffer.go
… vendor/github.com/parquet-go/parquet-go/column_chunk.go
… vendor/github.com/parquet-go/parquet-go/column_index.go
… vendor/github.com/parquet-go/parquet-go/column_index_be.go
… vendor/github.com/parquet-go/parquet-go/column_index_le.go
… vendor/github.com/parquet-go/parquet-go/config.go
… vendor/github.com/parquet-go/parquet-go/convert.go
… vendor/github.com/parquet-go/parquet-go/dedupe.go
… vendor/github.com/parquet-go/parquet-go/encoding/thrift/encode.go
… vendor/github.com/parquet-go/parquet-go/errors.go
… vendor/github.com/parquet-go/parquet-go/file.go
… vendor/github.com/parquet-go/parquet-go/internal/bytealg/count_amd64.s
… vendor/github.com/parquet-go/parquet-go/merge.go
… vendor/github.com/parquet-go/parquet-go/multi_row_group.go
… vendor/github.com/parquet-go/parquet-go/node.go
… vendor/github.com/parquet-go/parquet-go/offset_index.go
… vendor/github.com/parquet-go/parquet-go/page.go
… vendor/github.com/parquet-go/parquet-go/reader.go
… vendor/github.com/parquet-go/parquet-go/row.go
… vendor/github.com/parquet-go/parquet-go/row_group.go
… vendor/github.com/parquet-go/parquet-go/schema.go