**What this PR does**:

Updates parquet-go to the latest commit. This included a few breaking changes so code has been modified in Tempo. 

Benches are mixed in an interesting way. I have not dug into details but there was a regression on all queries that failed to match anything and nice improvements on a lot of the queries that found matches. The "no match" searches basically do nothing except repeatedly open columns so perhaps additional work was done there to speed up other parts of parquet-go?

<details>
<summary>benches</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: Apple M3 Pro
                                                    │ before.txt  │              after.txt              │
                                                    │   sec/op    │   sec/op     vs base                │
BackendBlockTraceQL/spanAttValMatch-11                97.19m ± 1%   83.61m ± 1%  -13.98% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              4.950m ± 1%   5.595m ± 1%  +13.03% (p=0.002 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          74.22m ± 0%   60.02m ± 0%  -19.13% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        5.060m ± 3%   5.534m ± 1%   +9.37% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            421.1m ± 0%   403.5m ± 0%   -4.18% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          5.254m ± 0%   5.189m ± 1%   -1.23% (p=0.001 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      38.41m ± 0%   25.00m ± 0%  -34.90% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   5.065m ± 0%   5.345m ± 0%   +5.52% (p=0.000 n=10)
BackendBlockTraceQL/traceOrMatch-11                   240.7m ± 0%   239.7m ± 0%   -0.42% (p=0.023 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 240.6m ± 0%   240.4m ± 1%        ~ (p=0.579 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                180.9m ± 0%   182.0m ± 0%   +0.60% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          5.107m ± 0%   5.070m ± 2%   -0.72% (p=0.019 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           149.6m ± 0%   148.6m ± 0%   -0.70% (p=0.000 n=10)
BackendBlockTraceQL/count-11                          343.4m ± 0%   332.2m ± 0%   -3.24% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         431.3m ± 2%   427.5m ± 2%   -0.88% (p=0.043 n=10)
BackendBlockTraceQL/||-11                             166.8m ± 3%   153.8m ± 0%   -7.78% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                          30.23m ± 0%   25.60m ± 0%  -15.33% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        5.029m ± 2%   5.105m ± 1%        ~ (p=0.063 n=10)
BackendBlockTraceQL/select-11                         5.075m ± 1%   5.102m ± 1%   +0.53% (p=0.011 n=10)
geomean                                               43.20m        41.28m        -4.44%

                                                    │  before.txt  │              after.txt               │
                                                    │     B/s      │     B/s       vs base                │
BackendBlockTraceQL/spanAttValMatch-11                230.4Mi ± 1%   267.9Mi ± 1%  +16.24% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              341.4Mi ± 1%   302.0Mi ± 1%  -11.53% (p=0.002 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          313.9Mi ± 0%   388.1Mi ± 0%  +23.66% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        489.6Mi ± 3%   447.6Mi ± 1%   -8.57% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            52.19Mi ± 0%   54.46Mi ± 0%   +4.37% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          171.2Mi ± 0%   173.3Mi ± 1%   +1.24% (p=0.001 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      573.3Mi ± 0%   880.7Mi ± 0%  +53.62% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   184.3Mi ± 0%   174.7Mi ± 0%   -5.24% (p=0.000 n=10)
BackendBlockTraceQL/traceOrMatch-11                   6.943Mi ± 0%   6.971Mi ± 0%   +0.41% (p=0.016 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 6.948Mi ± 0%   6.952Mi ± 1%        ~ (p=0.580 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                10.90Mi ± 0%   10.83Mi ± 0%   -0.57% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          174.8Mi ± 0%   176.1Mi ± 2%   +0.73% (p=0.019 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           18.44Mi ± 0%   18.57Mi ± 0%   +0.70% (p=0.000 n=10)
BackendBlockTraceQL/count-11                          63.97Mi ± 0%   66.12Mi ± 0%   +3.35% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         12.66Mi ± 2%   12.77Mi ± 2%   +0.90% (p=0.045 n=10)
BackendBlockTraceQL/||-11                             132.3Mi ± 3%   143.5Mi ± 0%   +8.44% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                          706.7Mi ± 0%   834.7Mi ± 0%  +18.11% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        179.0Mi ± 2%   176.3Mi ± 1%        ~ (p=0.063 n=10)
BackendBlockTraceQL/select-11                         177.3Mi ± 1%   176.4Mi ± 1%   -0.53% (p=0.011 n=10)
geomean                                               96.05Mi        100.5Mi        +4.65%

                                                    │ before.txt  │              after.txt               │
                                                    │  MB_io/op   │  MB_io/op    vs base                 │
BackendBlockTraceQL/spanAttValMatch-11                 23.48 ± 0%    23.48 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttValNoMatch-11               1.772 ± 0%    1.772 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicMatch-11           24.43 ± 0%    24.43 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11         2.598 ± 0%    2.598 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValMatch-11             23.04 ± 0%    23.04 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValNoMatch-11          943.2m ± 0%   943.2m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch-11       23.09 ± 0%    23.09 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   979.0m ± 0%   979.0m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/traceOrMatch-11                    1.753 ± 0%    1.753 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/traceOrNoMatch-11                  1.753 ± 0%    1.753 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValNoMatch-11                 2.067 ± 0%    2.067 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchAnd-11          936.1m ± 0%   936.1m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchOr-11            2.893 ± 0%    2.893 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/count-11                           23.03 ± 0%    23.03 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/struct-11                          5.726 ± 0%    5.726 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/||-11                              23.14 ± 0%    23.14 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixed-11                           22.40 ± 0%    22.40 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/complex-11                        943.7m ± 0%   943.7m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/select-11                         943.7m ± 0%   943.7m ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                4.351         4.351       +0.00%
¹ all samples are equal

                                                    │  before.txt   │               after.txt                │
                                                    │     B/op      │      B/op       vs base                │
BackendBlockTraceQL/spanAttValMatch-11                43.81Mi ±  1%    44.88Mi ±  1%   +2.44% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              5.623Mi ±  2%    6.688Mi ±  2%  +18.94% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          38.45Mi ±  1%    39.63Mi ±  1%   +3.08% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        5.694Mi ±  1%    7.068Mi ±  1%  +24.12% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            564.6Mi ±  0%    576.3Mi ±  0%   +2.07% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          5.259Mi ±  1%    5.264Mi ±  3%        ~ (p=0.579 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      9.845Mi ±  2%   10.534Mi ±  1%   +7.00% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   5.690Mi ±  1%    6.765Mi ±  1%  +18.89% (p=0.000 n=10)
BackendBlockTraceQL/traceOrMatch-11                   9.464Mi ± 22%    9.262Mi ± 18%        ~ (p=0.853 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 9.356Mi ± 22%    9.439Mi ± 15%        ~ (p=0.579 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                6.474Mi ± 10%    8.009Mi ±  9%  +23.71% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          5.573Mi ±  1%    5.784Mi ±  2%   +3.80% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           6.314Mi ± 10%    7.604Mi ±  4%  +20.44% (p=0.000 n=10)
BackendBlockTraceQL/count-11                          408.0Mi ±  1%    418.8Mi ±  0%   +2.66% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         14.16Mi ±  8%    13.61Mi ± 30%        ~ (p=0.853 n=10)
BackendBlockTraceQL/||-11                             16.08Mi ±  2%    16.86Mi ±  5%   +4.89% (p=0.019 n=10)
BackendBlockTraceQL/mixed-11                          6.404Mi ±  3%    6.995Mi ±  3%   +9.24% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        5.533Mi ±  3%    5.485Mi ±  3%        ~ (p=0.796 n=10)
BackendBlockTraceQL/select-11                         5.524Mi ±  2%    5.518Mi ±  2%        ~ (p=0.631 n=10)
geomean                                               13.57Mi          14.48Mi         +6.76%

                                                    │ before.txt  │             after.txt              │
                                                    │  allocs/op  │  allocs/op   vs base               │
BackendBlockTraceQL/spanAttValMatch-11                498.2k ± 0%   498.4k ± 0%  +0.06% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11              79.32k ± 0%   79.46k ± 0%  +0.18% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11          287.1k ± 0%   287.4k ± 0%  +0.10% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11        79.28k ± 0%   79.42k ± 0%  +0.19% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-11            3.444M ± 0%   3.445M ± 0%  +0.02% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11          79.36k ± 0%   79.46k ± 0%  +0.13% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11      125.3k ± 0%   125.7k ± 0%  +0.27% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11   79.31k ± 0%   79.46k ± 0%  +0.18% (p=0.000 n=10)
BackendBlockTraceQL/traceOrMatch-11                   86.17k ± 1%   86.35k ± 1%  +0.21% (p=0.011 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                 86.14k ± 0%   86.34k ± 0%  +0.23% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                80.08k ± 0%   80.29k ± 0%  +0.26% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11          79.36k ± 0%   79.46k ± 0%  +0.13% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11           80.09k ± 0%   80.27k ± 0%  +0.22% (p=0.000 n=10)
BackendBlockTraceQL/count-11                          1.985M ± 0%   1.985M ± 0%  +0.03% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                         93.95k ± 6%   94.13k ± 3%       ~ (p=0.684 n=10)
BackendBlockTraceQL/||-11                             161.1k ± 0%   161.5k ± 0%  +0.23% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                          87.48k ± 0%   87.84k ± 0%  +0.40% (p=0.000 n=10)
BackendBlockTraceQL/complex-11                        79.70k ± 0%   79.81k ± 0%  +0.14% (p=0.000 n=10)
BackendBlockTraceQL/select-11                         79.58k ± 0%   79.69k ± 0%  +0.15% (p=0.000 n=10)
geomean                                               147.1k        147.4k       +0.18%
```

</details>