@joe-elliott 2025-02-04
I'm not sure, but I believe it's one of these two:

https://github.com/parquet-go/parquet-go/pull/229
https://github.com/parquet-go/parquet-go/pull/227